requirements-api.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.20
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
safetensors>=0.4.0
